Universum.Container.Reexport

Description

This module reexports all container related stuff from Prelude.

Documentation

class Hashable a where

The class of types that can be converted to a hash value.

Minimal implementation: hashWithSalt.

Minimal complete definition

Nothing

Methods

hashWithSalt :: Int -> a -> Int infixl 0

Return a hash value for the argument, using the given salt.

The general contract of hashWithSalt is:

If two values are equal according to the == method, then applying the hashWithSalt method on each of the two values must produce the same integer result if the same salt is used in each case.

It is not required that if two values are unequal according to the == method, then applying the hashWithSalt method on each of the two values must produce distinct integer results. However, the programmer should be aware that producing distinct integer results for unequal values may improve the performance of hashing-based data structures.

This method can be used to compute different hash values for the same input by providing a different salt in each application of the method. This implies that any instance that defines hashWithSalt must make use of the salt in its implementation.

data HashMap k v

A map from keys to values. A map cannot contain duplicate keys; each key can map to at most one value.

data HashSet a

A set of values. A set cannot contain duplicate values.

data IntMap a

A map of integers to values a.

data IntSet

A set of integers.

data Map k a

A Map from keys k to values a.

data Seq a

General-purpose finite sequences.

data Set a

A set of values a.

fst :: (a, b) -> a

Extract the first component of a pair.

snd :: (a, b) -> b

Extract the second component of a pair.

swap :: (a, b) -> (b, a)

Swap the components of a pair.

uncurry :: (a -> b -> c) -> (a, b) -> c

uncurry converts a curried function to a function on pairs.

Examples

Expand

>>> uncurry (+) (1,2)
3

>>> uncurry ($) (show, 1)
"1"

>>> map (uncurry max) [(1,2), (3,4), (6,8)]
[2,4,8]

curry :: ((a, b) -> c) -> a -> b -> c

curry converts an uncurried function to a curried function.

Examples

Expand

>>> curry fst 1 2
1

data Vector a

Boxed vectors, supporting efficient slicing.