Universum.Debug

Description

Functions for debugging. If you left these functions in your code then warning is generated to remind you about left usages. Also, some functions (and data types) are convenient for prototyping.

Documentation

data Undefined

Warning: Undefined type remains in code

Similar to undefined but data type.

Constructors

Undefined

Warning: Undefined type remains in code

error :: forall (r :: RuntimeRep). forall (a :: TYPE r). HasCallStack => Text -> a

error that takes Text as an argument.

trace :: Text -> a -> a

Warning: trace remains in code

Version of trace that leaves a warning and takes Text.

traceM :: Monad m => Text -> m ()

Warning: traceM remains in code

Version of traceM that leaves a warning and takes Text.

traceId :: Text -> Text

Warning: traceId remains in code

Version of traceId that leaves a warning.

traceIdWith :: (a -> Text) -> a -> a

Warning: traceIdWith remains in code

Version of traceId that leaves a warning. Useful to tag printed data, for instance:

traceIdWith (x -> "My data: " <> show x) (veryLargeExpression)

This is especially useful with custom formatters:

traceIdWith (x -> "My data: " <> pretty x) (veryLargeExpression)

traceShow :: Show a => a -> b -> b

Warning: traceShow remains in code

Version of traceShow that leaves a warning.

traceShowId :: Show a => a -> a

Warning: traceShowId remains in code

Version of traceShowId that leaves a warning.

traceShowIdWith :: Show s => (a -> s) -> a -> a

Warning: traceShowIdWith remains in code

Version of traceShowId that leaves a warning. Useful to tag printed data, for instance:

traceShowIdWith ("My data: ", ) (veryLargeExpression)

traceShowM :: (Show a, Monad m) => a -> m ()

Warning: traceShowM remains in code

Version of traceShowM that leaves a warning.

undefined :: forall (r :: RuntimeRep). forall (a :: TYPE r). HasCallStack => a

Warning: undefined function remains in code (or use error)

undefined that leaves a warning in code on every usage.