Universum.Lifted.Concurrent

Description

Concurrency useful and common functions.

MVar

data MVar a

An MVar (pronounced "em-var") is a synchronising variable, used for communication between concurrent threads. It can be thought of as a box, which may be empty or full.

newEmptyMVar :: MonadIO m => m (MVar a)

Lifted to MonadIO version of newEmptyMVar.

newMVar :: MonadIO m => a -> m (MVar a)

Lifted to MonadIO version of newMVar.

putMVar :: MonadIO m => MVar a -> a -> m ()

Lifted to MonadIO version of putMVar.

readMVar :: MonadIO m => MVar a -> m a

Lifted to MonadIO version of readMVar.

swapMVar :: MonadIO m => MVar a -> a -> m a

Lifted to MonadIO version of swapMVar.

takeMVar :: MonadIO m => MVar a -> m a

Lifted to MonadIO version of takeMVar.

tryPutMVar :: MonadIO m => MVar a -> a -> m Bool

Lifted to MonadIO version of tryPutMVar.

tryReadMVar :: MonadIO m => MVar a -> m (Maybe a)

Lifted to MonadIO version of tryReadMVar.

tryTakeMVar :: MonadIO m => MVar a -> m (Maybe a)

Lifted to MonadIO version of tryTakeMVar.

STM

data STM a

A monad supporting atomic memory transactions.

data TVar a

Shared memory locations that support atomic memory transactions.

atomically :: MonadIO m => STM a -> m a

Lifted to MonadIO version of atomically.

newTVarIO :: MonadIO m => a -> m (TVar a)

Lifted to MonadIO version of newTVarIO.

readTVarIO :: MonadIO m => TVar a -> m a

Lifted to MonadIO version of readTVarIO.

modifyTVar' :: TVar a -> (a -> a) -> STM ()

Strict version of modifyTVar.

Since: stm-2.3

newTVar :: a -> STM (TVar a)

Create a new TVar holding a value supplied

readTVar :: TVar a -> STM a

Return the current value stored in a TVar.

writeTVar :: TVar a -> a -> STM ()

Write the supplied value into a TVar.