Universum.List.Safe

Description

This module contains safe functions to work with list type (mostly with NonEmpty).

Documentation

uncons :: [a] -> Maybe (a, [a])

Destructuring list into its head and tail if possible. This function is total.

>>> uncons []
Nothing
>>> uncons [1..5]
Just (1,[2,3,4,5])
>>> uncons (5 : [1..5]) >>= \(f, l) -> pure $ f == length l
Just True

whenNotNull :: Applicative f => [a] -> (NonEmpty a -> f ()) -> f ()

Performs given action over NonEmpty list if given list is non empty.

>>> whenNotNull [] $ \(b :| _) -> print (not b)
>>> whenNotNull [False,True] $ \(b :| _) -> print (not b)
True

whenNotNullM :: Monad m => m [a] -> (NonEmpty a -> m ()) -> m ()

Monadic version of whenNotNull.