Universum.Monad.Either

Description

Utilites to work with Either data type.

Documentation

fromLeft :: a -> Either a b -> a

Extracts value from Left or return given default value.

>>> fromLeft 0 (Left 3)
3
>>> fromLeft 0 (Right 5)
0

fromRight :: b -> Either a b -> b

Extracts value from Right or return given default value.

>>> fromRight 0 (Left 3)
0
>>> fromRight 0 (Right 5)
5

maybeToLeft :: r -> Maybe l -> Either l r

Maps Maybe to Either wrapping default value into Right.

>>> maybeToLeft True (Just "aba")
Left "aba"
>>> maybeToLeft True Nothing
Right True

maybeToRight :: l -> Maybe r -> Either l r

Maps Maybe to Either wrapping default value into Left.

>>> maybeToRight True (Just "aba")
Right "aba"
>>> maybeToRight True Nothing
Left True

leftToMaybe :: Either l r -> Maybe l

Maps left part of Either to Maybe.

>>> leftToMaybe (Left True)
Just True
>>> leftToMaybe (Right "aba")
Nothing

rightToMaybe :: Either l r -> Maybe r

Maps right part of Either to Maybe.

>>> rightToMaybe (Left True)
Nothing
>>> rightToMaybe (Right "aba")
Just "aba"

whenLeft :: Applicative f => Either l r -> (l -> f ()) -> f ()

Applies given action to Either content if Left is given.

whenLeftM :: Monad m => m (Either l r) -> (l -> m ()) -> m ()

Monadic version of whenLeft.

whenRight :: Applicative f => Either l r -> (r -> f ()) -> f ()

Applies given action to Either content if Right is given.

whenRightM :: Monad m => m (Either l r) -> (r -> m ()) -> m ()

Monadic version of whenRight.