Universum.Print.Internal

Description

You may import this module to define your own, custom instances of Print. Be advised however that this module is an internal API and may be subject to change even for minor version increments.

Documentation

class Print a where

Support class to overload writing of string like values.

Methods

hPutStr :: Handle -> a -> IO ()

hPutStrLn :: Handle -> a -> IO ()