Universum.Unsafe

Description

Unsafe functions to work with lists and Maybe. Sometimes unavoidable but better don't use them. This module is intended to be imported qualified and it's not even included in default prelude exports.

import qualified Universum.Unsafe as Unsafe

foo :: [a] -> a
foo = Unsafe.head

Documentation

head :: [a] -> a

Extract the first element of a list, which must be non-empty.

tail :: [a] -> [a]

Extract the elements after the head of a list, which must be non-empty.

init :: [a] -> [a]

Return all the elements of a list except the last one. The list must be non-empty.

last :: [a] -> a

Extract the last element of a list, which must be finite and non-empty.

at :: Int -> [a] -> a

Similar to !! but with flipped arguments.

(!!) :: [a] -> Int -> a infixl 9

List index (subscript) operator, starting from 0. It is an instance of the more general genericIndex, which takes an index of any integral type.

fromJust :: Maybe a -> a

The fromJust function extracts the element out of a Just and throws an error if its argument is Nothing.

Examples

Expand

Basic usage:

>>> fromJust (Just 1)
1

>>> 2 * (fromJust (Just 10))
20

>>> 2 * (fromJust Nothing)
*** Exception: Maybe.fromJust: Nothing